refactor partner models to include ai21
